+++ app/src/main/java/com/ikvych/cocktail/data/entity/DrinkApiResponse.java
@@ -1,11 +1,10 @@
-package com.ikvych.cocktail.model;
+package com.ikvych.cocktail.data.entity;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.os.Parcelable.Creator;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
@@ -30,7 +29,7 @@ public class DrinkApiResponse implements Parcelable {
     };
 
     protected DrinkApiResponse(Parcel in) {
-        in.readList(this.drinks, (com.ikvych.cocktail.model.Drink.class.getClassLoader()));
+        in.readList(this.drinks, (Drink.class.getClassLoader()));
     }
 
     public DrinkApiResponse() {
