+++ app/src/main/java/com/ikvych/cocktail/ui/activity/MainActivity.java
@@ -1,13 +1,14 @@
-package com.ikvych.cocktail.view;
+package com.ikvych.cocktail.ui.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
 
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.ikvych.cocktail.R;
-import com.ikvych.cocktail.adapter.DrinkAdapter;
-import com.ikvych.cocktail.model.Drink;
-import com.ikvych.cocktail.util.ActivityUtil;
+import com.ikvych.cocktail.ect.adapter.DrinkAdapter;
+import com.ikvych.cocktail.data.entity.Drink;
+import com.ikvych.cocktail.ui.base.BaseActivity;
+import com.ikvych.cocktail.ect.util.ActivityUtil;
 import com.ikvych.cocktail.viewmodel.MainActivityViewModel;
 
 import java.util.List;
