+++ app/src/main/java/com/ikvych/cocktail/viewmodel/SearchActivityViewModel.java
@@ -5,13 +5,14 @@ import android.app.Application;
 import androidx.annotation.NonNull;
 import androidx.lifecycle.MutableLiveData;
 
-import com.ikvych.cocktail.model.Drink;
-import com.ikvych.cocktail.repository.DrinkApiRepository;
-import com.ikvych.cocktail.repository.impl.DrinkApiRepositoryImpl;
+import com.ikvych.cocktail.data.entity.Drink;
+import com.ikvych.cocktail.data.repository.base.DrinkApiRepository;
+import com.ikvych.cocktail.data.repository.DrinkApiRepositoryImpl;
+import com.ikvych.cocktail.viewmodel.base.BaseViewModel;
 
 import java.util.List;
 
-public class SearchActivityViewModel extends ActivityViewModel {
+public class SearchActivityViewModel extends BaseViewModel {
 
     private DrinkApiRepository drinkApiRepository;
