+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.SearchActivity">
+        tools:context=".ui.activity.SearchActivity">
 
         <LinearLayout
             android:id="@+id/search_container"
